- added global login func for all torrent providers (all private trackers need cookies) - added global download func (same, downloading also requires cookies)
These 2 functions were necessary to allow the user to download other found releases at a later time.
In the current version this doesn't work because it goes to the url directly, without any cookies
which causes it to download the login page instead of a torrent.

All 3 providers now properly use the cache and have much better error handling.